[Step3:refactor] update함수 수정 및 TestCode 추가
