+++ src/main/java/gift/repository/ProductRepository.java
@@ -47,10 +47,10 @@ public class ProductRepository {
     return product;
   }
 
-  public void update(Product oldProduct, Product updateProduct) {
+  public void update(Product updateProduct) {
     String sql = "update product set name = ?, price = ?, imageurl = ? where id = ?";
     jdbcTemplate.update(sql, updateProduct.getName(), updateProduct.getPrice(),
-        updateProduct.getImageUrl(), oldProduct.getId());
+        updateProduct.getImageUrl(), updateProduct.getId());
   }
 
   public void delete(Long id) {
