+++ src/main/resources/data.sql
@@ -1,0 +1,9 @@
+INSERT INTO product (name, price, imageUrl)
+VALUES ('아이스 카페 아메리카노 T',
+        600,
+        'https://st.kakaocdn.net/product/gift/product/20231010111814_9a667f9eccc943648797925498bdd8a3.jpg');
+
+INSERT INTO product (name, price, imageUrl)
+VALUES ('코코아 파우더',
+        7600,
+        'https://encrypted-tbn3.gstatic.com/shopping?q=tbn:ANd9GcRWDlZJBZiXDT8JlPp8iS5eeyJDdsISVl9lBSTXgcLhUAjCjU-qttjbpgg8-JPRHMWWOKEuGqlQWTqIvg_Hw_dAaEdZYPM-JF5z7KMOGPGkpGIj3z9WZMJG');
